5 lesson FIX: migrations foreign crash

+++ database/migrations/2021_01_16_223103_create_blog_posts_table.php
@@ -15,8 +15,8 @@ class CreateBlogPostsTable extends Migration
     {
         Schema::create('blog_posts', function (Blueprint $table) {
             $table->id();
-            $table->integer('category_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('category_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->string('slug')->unique();
             $table->string('title');
             $table->text('excerpt')->nullable();
